Refactor of code base to be more in line with pep guide
