+++ requirements.txt
@@ -1,0 +1,2 @@
+tifffile==2023.4.12
+numpy==1.24.3
